Moving all the threading operations to it's own file
